Bugfix from feedback

+++ internal/domain/transaction/transaction.go
@@ -16,7 +16,7 @@ type Transaction struct {
 	ReferenceCode     string               `json:"referenceCode"`
 	UserId            string               `json:"userId"`
 	CreatedTime       string               `json:"createdTime"`
-	Total             int64                `json:"total"`
+	Total             float64                `json:"total"`
 	TransactionDetail []*TransactionDetail `json:"transactionDetail"`
 }
 
@@ -27,8 +27,8 @@ type TransactionDetail struct {
 	ProductID     string  `json:"productId"`
 	BuyPrice      float64 `json:"buyPrice"`
 	SellPrice     float64 `json:"sellPrice"`
-	Quantity      int64   `json:"quantity"`
-	BuyQuantity   int64   `json:"buyQuantity"`
+	Quantity      float64   `json:"quantity"`
+	BuyQuantity   float64   `json:"buyQuantity"`
 }
 
 type TransactionStatus struct {
@@ -44,7 +44,7 @@ type TransactionStatus struct {
 	UserId            string                     `json:"userId"`
 	UserName          string                     `json:"userName"`
 	CreatedTime       string                     `json:"createdTime"`
-	Total             int64                      `json:"total"`
+	Total             float64                      `json:"total"`
 	TransactionDetail []*TransactionStatusDetail `json:"transactionDetail"`
 }
 
@@ -57,6 +57,6 @@ type TransactionStatusDetail struct {
 	ProductName   string  `json:"productName"`
 	BuyPrice      float64 `json:"buyPrice"`
 	SellPrice     float64 `json:"sellPrice"`
-	Quantity      int64   `json:"quantity"`
-	BuyQuantity   int64   `json:"buyQuantity"`
+	Quantity      float64   `json:"quantity"`
+	BuyQuantity   float64   `json:"buyQuantity"`
 }
